requirements.txt
streamlit
fastapi
openai
sqlalchemy
uvicorn
pandas
python-dotenv
